feat(components): added if to check if user is owner - needed for admin

+++ src/components/LoactionList/LocationList.tsx
@@ -14,7 +14,11 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Link as RouterLink } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { selectUserLocations, selectUserToken } from '../../store/selectors'
+import {
+  selectUserLocations,
+  selectUserToken,
+  selectUserId,
+} from '../../store/selectors'
 import { deletePlace } from '../../utils/CreateNewPlace'
 import { initUserData } from '../../utils/initUserData'
 
@@ -28,38 +32,41 @@ const LoactionListItem = withTheme(styled(ListItem)`
 export const LocationList = () => {
   const locations = useSelector(selectUserLocations())
   const token = useSelector(selectUserToken())
+  const userIDstring = useSelector(selectUserId())
+  const userID: number = +userIDstring
   const dispatch = useDispatch()
 
   return (
     <List>
       {locations?.map((location) => {
-        return (
-          <LoactionListItem button key={location.id}>
-            <Box display="flex" justifyContent="space-between" width="100%">
-              <Box display="flex" flexDirection="column">
-                <ListItemText primary={location.name} />
-                <ListItemText secondary={location.description} />
+        if (location.user_id === userID)
+          return (
+            <LoactionListItem button key={location.id}>
+              <Box display="flex" justifyContent="space-between" width="100%">
+                <Box display="flex" flexDirection="column">
+                  <ListItemText primary={location.name} />
+                  <ListItemText secondary={location.description} />
+                </Box>
+                <Box display="flex" flexDirection="column">
+                  <IconButton
+                    component={RouterLink}
+                    to={`/neuer_ort/edit/:${location.id}`}
+                  >
+                    <EditIcon />
+                  </IconButton>
+                  <IconButton
+                    onClick={async () => {
+                      const response = await deletePlace(token, location.id)
+                      if (response.status && response.status === 204)
+                        await initUserData(token, dispatch)
+                    }}
+                  >
+                    <DeleteIcon />
+                  </IconButton>
+                </Box>
               </Box>
-              <Box display="flex" flexDirection="column">
-                <IconButton
-                  component={RouterLink}
-                  to={`/neuer_ort/edit/:${location.id}`}
-                >
-                  <EditIcon />
-                </IconButton>
-                <IconButton
-                  onClick={async () => {
-                    const response = await deletePlace(token, location.id)
-                    if (response.status && response.status === 204)
-                      await initUserData(token, dispatch)
-                  }}
-                >
-                  <DeleteIcon />
-                </IconButton>
-              </Box>
-            </Box>
-          </LoactionListItem>
-        )
+            </LoactionListItem>
+          )
       })}
     </List>
   )
